PIG-361.  Created a new type PigNullableWritable that extends WritableComparable.  In addition to a null byte this type includes an index byte, so that when types are used in a join, this byte tracks which input the record came from.  Changed all NullableXWritable types to extend PigNullableWritable.  Created new NullableTuple that also extends PigNullableWritable.  Wrote the comparator of PigNullableWritable so that it ignores the index byte accept when the null byte is true.  This results in the desired cogroup semantics where nulls of a given input coallesce but nulls across inputs do not.
Changed all map reduce jobs so that keys are always a PigNullableWritable.  This was necessary so that the index could be used in sorting.  In the future this
will allow join optimizations by having join tuples sorted by input source.  Also changed so that all values are NullableTuples.  This replaces the
IndexedTuple.

Fixed a bug in the SortPartitioner where it was always assuming that inputs were tuples, which is no longer the case.



git-svn-id: https://svn.apache.org/repos/asf/incubator/pig/branches/types@697229 13f79535-47bb-0310-9956-ffa450edef68