Update pig.cmd
hadoop-config.cmd is not present in the bin folder of the HADOOP_HOME instead it is present in the libexec folder. I was getting exception   
when I ran pig.cmd -x local for the local mode I was getting the exception and hen I changed as I forked then it started working.